day_2/homework_1_setting_up_vector_db/pyproject.toml
[tool.poetry]
name = "homework_setting_up_vector_db"
version = "0.1.0"
description = ""
authors = ["Launch School"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
openai = "1.79.0"
psycopg2 = "2.9.10"
python-dotenv = "1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
